+++ packages/vue-stack/src/constants.ts
@@ -1,1 +1,3 @@
 export const V_STACK_ACTIVATED_ATTR = 'data-v-stack-activated';
+
+export const V_STACK_TYPE_ATTR = 'data-v-stack-type';