+++ packages/vite-plugin-vui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vite-plugin-vui
 
+## 2.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.20.3
+
 ## 2.2.2
 
 ### Patch Changes